+++ src/constants/index.js
@@ -4,3 +4,4 @@ export { default as changePasswordErrorCodes } from './changePasswordErrorCodes'
 export { default as defaultErrors } from './defaultErrors';
 export { default as packageName } from './packageName';
 export { default as delimiters } from './delimiters';
+export * from './shapes';